Merge dev to master

+++ src/dao/UserDao.ts
@@ -33,7 +33,7 @@ export default class UserDao implements IDao<IUserDatabaseObject> {
   }
 
   /**
-   * Finds a single user by its username
+   * Finds a single user by its username.
    *
    * @param {string} username Username
    * @returns {Promise<IUserDatabaseObject>} User
@@ -43,6 +43,20 @@ export default class UserDao implements IDao<IUserDatabaseObject> {
     return this.knex("users")
       .select()
       .where({ username })
+      .first();
+  }
+
+  /**
+   * Finds a single user by its email address.
+   *
+   * @param {string} email Email address
+   * @returns {Promise<IUserDatabaseObject>} User
+   * @memberof UserDao
+   */
+  public findByEmail(email: string): Promise<IUserDatabaseObject> {
+    return this.knex("users")
+      .select()
+      .where({ email })
       .first();
   }
 
@@ -114,10 +128,13 @@ export default class UserDao implements IDao<IUserDatabaseObject> {
    * @returns {Promise<IUserDatabaseObject[]>}
    * @memberof UserDao
    */
-  public findAll(fields?: string[], conditions?: string[]): Promise<IUserDatabaseObject[]> {
+  public findAll(
+    fields?: string[],
+    conditions?: string[]
+  ): Promise<IUserDatabaseObject[]> {
     if (fields) {
       const queryString: string = fields.join("`, ");
-      let query: any = this.knex("users").select(fields);
+      let query: Knex.QueryBuilder = this.knex("users").select(fields);
 
       if (queryString.indexOf("Payment.")) {
         query.leftOuterJoin("payments", "users.id", "payments.payer_id");
@@ -162,20 +179,33 @@ export default class UserDao implements IDao<IUserDatabaseObject> {
    * @memberof UserDao
    */
   public remove(id: number): Promise<boolean> {
-    return this.knex("users")
+    // First, delete consents
+    return this.knex("privacy_policy_consent_data")
       .delete()
-      .where({ id });
+      .where({ user_id: id })
+      .then<boolean>((result: boolean) => {
+        return this.knex("users")
+          .delete()
+          .where({ id });
+      });
   }
 
   /**
-   * Updates a singler user.
+   * Updates a single user.
    *
    * @param {number} entityId User id
    * @param {IUserDatabaseObject} entity Entity
-   * @returns {Promise<boolean>}
+   * @returns {Promise<number[]>} Affected rows
    * @memberof UserDao
    */
-  public update(entityId: number, entity: IUserDatabaseObject): Promise<boolean> {
+  public update(
+    entityId: number,
+    entity: IUserDatabaseObject
+  ): Promise<number> {
+    if (entity.created) {
+      delete entity.created;
+    }
+    entity.modified = new Date();
     return this.knex("users")
       .update(entity)
       .where({ id: entityId });
@@ -185,10 +215,16 @@ export default class UserDao implements IDao<IUserDatabaseObject> {
    * Saves a single user.
    *
    * @param {IUserDatabaseObject} entity
-   * @returns {Promise<number[]>}
+   * @returns {Promise<number[]>} Inserted ID(s)
    * @memberof UserDao
    */
   public save(entity: IUserDatabaseObject): Promise<number[]> {
+    // Delete id because it's auto-assigned
+    if (entity.id) {
+      delete entity.id;
+    }
+    entity.created = new Date();
+    entity.modified = new Date();
     return this.knex("users").insert(entity);
   }
 }
